Remove paused from base check schema, rename check to HTTP check
The "paused" parameter is inconsistently implemented - it's an option
for create/modify but not for reading. Not implementing for now - may
add back if there is sufficient demand for it.

Also, have decided to break up each check type up into a resource. The
SDK is unified, so this is a trivial task, however may allow for better
and more understandable Terraform code.